Added specific color values
Changed Board, ScoreCounter, Item, and  ItemDatabase scripts.
Item now has a variable containing the element. This can be set in Unity (in a similar manner to the value of an item)
ItemDatabase initializes a list of possible elements before the scene loads.
ScoreCounter has a new function, add_score, which takes in the element type and value. It uses a dictionary to assign the score to a element and prints it.
Board now calls add_score to update the value and passes the element type additionally now.